+++ CHANGES.md
@@ -9,6 +9,19 @@
 - FIX
     - バグ修正
 
+## develop
+
+## 1.0.2
+
+- [CHANGE] `RTCVideoView`: `objectFit` プロパティに `fill` オプションを追加した
+
+- [FIX] `getUserMedia()`: `facingMode` に `user` を指定しても前面カメラに切り換わらない現象を修正する @kdxu
+
+## 1.0.1
+
+- [FIX] audio トラックが常に NO になってしまう問題を修正 @kdxu
+- [FIX] ドキュメントの typo 修正 @kdxu
+
 ## 1.0.0
 
 リリース